fix pause progress sync

+++ src/api/sessions/sync_open_session.rs
@@ -8,7 +8,7 @@ use serde_json::json;
 /// https://api.audiobookshelf.org/#sync-an-open-session
 
 // sync a session
-pub async fn sync_session(token: Option<&String>, session_id: &str, current_time: Option<u32>, time_listened: u64, server_address: String) -> Result<(), reqwest::Error> {
+pub async fn sync_session(token: Option<&String>, session_id: &str, current_time: Option<u32>, time_listened: u32, server_address: String) -> Result<(), reqwest::Error> {
     let client = Client::new();
 
     let params = json!({
